clangtidy

+++ csrc/multidevice/propagation.h
@@ -14,12 +14,12 @@
 
 namespace nvfuser {
 
-int numParallelIterDomains(const TensorView* tv);
+int64_t numParallelIterDomains(const TensorView* tv);
 
 template <typename R>
 TensorView* findMostParallelTensorView(const R& range) {
   TensorView* reference = nullptr;
-  int max_parallel_count = -1;
+  int64_t max_parallel_count = -1;
   for (TensorView* tv : range) {
     auto parallel_count = numParallelIterDomains(tv);
     if (parallel_count > max_parallel_count) {
